backend/pyproject.toml
[tool.poetry]
name = "aura-backend"
version = "0.1.0"
description = "AURA Backend - FastAPI service for healthcare AI platform"
authors = ["AURA Team"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
pydantic = {extras = ["email"], version = "^2.9.2"}
pydantic-settings = "^2.5.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.12"
openai = "^1.54.0"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
apscheduler = "^3.10.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


